% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{fmt_functions_test}
\alias{fmt_functions_test}
\title{Data frame used for testing fmt_* functions}
\format{a dataframe}
\usage{
fmt_functions_test
}
\description{
Data frame used for testing fmt_* functions
}
\keyword{datasets}
