\name{prettyfunctionprint}
\alias{prettyfunctionprint}
\title{This helper function creates nice output strings for locatr functions}
\usage{
  prettyfunctionprint(procName, anstr, flagb)
}
\arguments{
  \item{procName}{the name of a working function to return}

  \item{anstr}{the data as a string}

  \item{flagb}{a character string containing "binary",
  "paren" or "normal"}
}
\value{
  a nicely formatted string
}
\description{
  This helper function creates nice output strings for
  locatr functions
}

