% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_expressions_to_quosures}
\alias{string_expressions_to_quosures}
\title{Convert  range to a filtering quosure}
\usage{
string_expressions_to_quosures(string_expression, environ)
}
\arguments{
\item{string_expression}{a string expression}

\item{environ}{an  environment}
}
\description{
Convert an expression that represents a spreadsheet range to a quosure
that can be used to create a variable that returns TRUE if row-col
combination is in the range.
}
