\name{isitperm}
\alias{isitperm}
\title{The function that asks you before it tests every single function}
\usage{
  isitperm(data, answer)
}
\arguments{
  \item{data}{the data (in a list) that you want to test}

  \item{answer}{the answer you expect}
}
\description{
  This functions will probably not be in the final version
  because it's quite time consuming, but can be useful for
  debugging.
}

